The data is fetched only for divisible current minutes

+++ src/main/java/vento/weather_scraper/config/OpenWeatherMapConfig.java
@@ -15,7 +15,7 @@ public class OpenWeatherMapConfig {
     /**
      * Delay for scheduler in minutes. Configured in application properties.
      */
-    private Long delay;
+    private Integer delay;
 
     /**
      * Token for accessing the OpenWeatherMap API. Configured in application properties.
